pyproject.toml
[tool.poetry]
name = "digital-games-piracy-data"
version = "0.1.1"
description = "A project for gathering and cleaning data regarding video game piracy."
authors = ["Matheus Araujo (matheusalexandre.engcomp@gmail.com"]
license = "Copyright (C) 2024 Matheus Araujo"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.3.0"
numpy = "^1.14.3"
requests = "^2.31.0"
apache-beam = {extras = ["gcp"], version = "^2.53.0"}
re101 = "^0.4.0"
python-dotenv = "^1.0.1"
pymongo = "^4.6.1"
selenium = "^4.17.2"
beautifulsoup4 = "^4.12.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
